+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Unless otherwise specified, any version comparison below is the comparison of se
 - (API) Added APIs for the new `DebugShape` system. Available debug shape classes can be found under `org.allaymc.api.debugshape`,
   and add/remove methods for debug shape are available in `Dimension` class.
 - (API) Added method `World.getState` as the replacement for the removed `World.isRunning` method.
+- (API) Added `ServerSettings.NetworkSettings.raknetGlobalPacketLimit` to control the global packet limit for RakNet.
 
 ### Changed